@@ -52,18 +48,27 @@
                 <role>maintainer</role>
             </roles>
         </developer>
+        <developer>
+            <name>Ferdi Gueran</name>
+            <email>ferdi.gueran@deepshore.de</email>
+            <url>https://github.com/deepshore</url>
+            <roles>
+                <role>maintainer</role>
+            </roles>
+        </developer>
     </developers>
     <scm>
-        <connection>scm:git:https://github.com/jcustenborder/kafka-connect-xml-parent.git</connection>
-        <developerConnection>scm:git:git@github.com:jcustenborder/kafka-connect-xml-parent.git</developerConnection>
-        <url>https://github.com/jcustenborder/kafka-connect-xml-parent</url>
+        <connection>scm:git:https://github.com/deepshore/xjc-kafka-connect-plugin.git</connection>
+        <developerConnection>scm:git:git@github.com:deepshore/xjc-kafka-connect-plugin.git</developerConnection>
+        <url>https://github.com/deepshore/xjc-kafka-connect-plugin</url>
     </scm>
     <issueManagement>
         <system>github</system>
-        <url>https://github.com/jcustenborder/kafka-connect-xml-parent/issues</url>
+        <url>https://github.com/deepshore/xjc-kafka-connect-plugin/issues</url>
     </issueManagement>
     <properties>
-        <jaxb.version>2.3.1</jaxb.version>
+        <java.version>17</java.version>
+        <jaxb.version>4.0.0</jaxb.version>
         <slf4j.version>1.7.21</slf4j.version>
         <junit.version>5.0.0</junit.version>
         <guava.version>30.1.1-jre</guava.version>
