@@ -23,10 +23,10 @@ import org.junit.jupiter.api.TestFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBElement;
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.Unmarshaller;
+import jakarta.xml.bind.JAXBContext;
+import jakarta.xml.bind.JAXBElement;
+import jakarta.xml.bind.JAXBException;
+import jakarta.xml.bind.Unmarshaller;
 import javax.xml.datatype.XMLGregorianCalendar;
 import java.io.IOException;
 import java.io.InputStream;
@@ -84,6 +84,8 @@ public abstract class AbstractRoundTripTest<T extends Connectable> {
         .map(method -> dynamicTest(method.getName(), () -> {
           Object exp = method.invoke(expected);
           Object act = method.invoke(actual);
+
+
 
           if (byte[].class.equals(method.getReturnType())) {
             byte[] ebytes = (byte[]) exp;
